make SurfaceImageMixin for common code
